Extract `picfiletype` and `farptr` structs into separate module
